refactor: ScopeArticleItemData type 분리(#46)
