+++ src/shared/components/scopeArticleItem/constants.ts
@@ -1,8 +1,8 @@
+import type { IdeologyType } from './types';
+
 /**
  * ScopeArticleItem 컴포넌트의 상수
  */
-
-export type IdeologyType = 'progressive' | 'center' | 'conservative';
 
 export const IDEOLOGY_LABELS: Record<IdeologyType, string> = {
   progressive: '진보',
